Moved summary stat logic into its own module
