+++ client/index.jsx
@@ -4,6 +4,7 @@ import axios from 'axios';
 import Summary from './components/Summary.jsx';
 import ReviewList from './components/ReviewList.jsx';
 import styled from 'styled-components';
+import stats from './stats.js';
 
 var Modal = styled.div`
   display: none;
@@ -70,6 +71,7 @@ class App extends React.Component {
       ]
     }
 
+    this.stats = {};
     this.open = this.open.bind(this);
     this.close = this.close.bind(this);
   }
@@ -100,6 +102,7 @@ class App extends React.Component {
   }
 
   render() {
+    this.stats = stats.getStats(this.state.reviews);
     return (
       <div id="reviews">
         <button id="open" onClick={this.open}>Reviews</button>
@@ -108,7 +111,7 @@ class App extends React.Component {
             <button id="close" onClick={this.close}>&#10005;</button>
           </Sticky>
           <ModalContent>
-            <Summary reviews={this.state.reviews}/>
+            <Summary stats={this.stats}/>
             <ReviewList reviews={this.state.reviews} />
           </ModalContent>
         </Modal>
